Updating Ditloids

+++ catscripts/Language/Ditloids/Ditloids.sh
@@ -1,14 +1,30 @@
 #!/bin/bash
 
-egrep -i '\[\[[0-9]{1,}=[a-z]{1,1}\s|\[\[[0-9]{1,}= [a-z]{1,1}\s|\[\[[0-9]{1,} = [a-z]{1,1}\s|\[\[[0-9]{1,} =[a-z]{1,1}\s|ditloid' newpages.txt >> Ditloids.txt
+KEYWORDS_DITLOIDS="^[0-9]{1,}=[a-z]{1,1}\s|^[0-9]{1,}= [a-z]{1,1}\s|^[0-9]{1,} = [a-z]{1,1}\s|^[0-9]{1,} =[a-z]{1,1}\s|ditloid"
 
-DITLOIDS=`stat --print=%s Ditloids.txt`
-
-if [ $DITLOIDS -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="Ditloids.txt"
-  export CATNAME="Ditloids"
-  $CATEGORIZE
-fi
 
-rm Ditloids.txt
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Ditloids\n"
+  fi
+
+  DITLOIDS=`egrep -i "$KEYWORDS_DITLOIDS" newpages.txt`
+
+  if [ "$DITLOIDS" != "" ];
+  then
+    printf "$DITLOIDS" > Ditloids.txt
+    export CATFILE="Ditloids.txt"
+    export CATNAME="Ditloids"
+    $CATEGORIZE
+    rm Ditloids.txt
+    unset DITLOIDS
+  fi
+
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Completed Ditloids\n"
+  fi
+
+fi
